+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * `Kreait\Firebase\Messaging\CloudMessage` can now be created without a target. The existence 
   of a message target is now validated on send. This enables re-using a message for multiple 
   targets.
+* Improved reliability of discovering a ServiceAccount from environment variables. 
+  (thanks to [@Shifu33](https://github.com/Shifu33)) for helping to find and test this) 
 
 ## 4.21.1 - 2019-05-14